refactor : moved functions to separate file
