+++ src/components/CartProduct/CartProduct.jsx
@@ -1,17 +1,20 @@
-import { useCart } from "context/cart-context";
-import { useWishlist } from "context/wishlist-context";
+import { useCart, useWishlist, useAuth } from "context";
 import React from "react";
+import { removeFromCart } from "services/cart";
+import { useToast } from "hooks/useToast";
+import { incrementQuantity, decrementQuantity } from "services/cart";
+import { moveToWishlist } from "services/wishlist";
+import { useNavigate } from "react-router-dom";
 import "./CartProduct.css";
+import axios from "axios";
 
 export const CartProduct = () => {
-  const {
-    isDisable,
-    itemsAdded,
-    incrementQuantity,
-    decrementQuantity,
-    removeFromCart,
-  } = useCart();
-  const { moveToWishlist } = useWishlist();
+  const { itemsAdded } = useCart();
+  const { wishlistItems, setWishlistItems } = useWishlist();
+  const { setItemsAdded, setIsDisable } = useCart();
+  const { auth } = useAuth();
+  const { showToast } = useToast();
+  const navigate = useNavigate();
 
   return (
     <div>
@@ -31,16 +34,27 @@ export const CartProduct = () => {
                 <div className="cart-quantity">
                   <div className="cart-quantity-head">Quantity:</div>
                   <button
-                    disabled={isDisable}
                     className="cart-quantity-minus"
-                    onClick={() => decrementQuantity(product)}
+                    onClick={() =>
+                      decrementQuantity(
+                        product,
+                        auth,
+                        setItemsAdded,
+                        axios,
+                        setIsDisable,
+                        removeFromCart,
+                        showToast
+                      )
+                    }
                   >
                     -
                   </button>
                   <div className="cart-quantity-value">{qty}</div>
                   <button
                     className="cart-quantity-plus"
-                    onClick={() => incrementQuantity(product)}
+                    onClick={() =>
+                      incrementQuantity(product, auth, setItemsAdded, axios)
+                    }
                   >
                     +
                   </button>
@@ -48,13 +62,33 @@ export const CartProduct = () => {
                 <div className="cart-buttons">
                   <button
                     className="cart-button-remove"
-                    onClick={() => removeFromCart(product)}
+                    onClick={() =>
+                      removeFromCart(
+                        product,
+                        axios,
+                        auth,
+                        setItemsAdded,
+                        showToast
+                      )
+                    }
                   >
                     remove from cart
                   </button>
                   <button
                     className="cart-button-move"
-                    onClick={() => moveToWishlist(product)}
+                    onClick={() =>
+                      moveToWishlist(
+                        product,
+                        wishlistItems,
+                        removeFromCart,
+                        showToast,
+                        auth,
+                        navigate,
+                        axios,
+                        setWishlistItems,
+                        setItemsAdded
+                      )
+                    }
                   >
                     move to wishlist
                   </button>
