+++ src/components/Login/Login.jsx
@@ -5,6 +5,7 @@ import { useAuth } from "../../context/auth-context";
 import { useNavigate, Link, useLocation } from "react-router-dom";
 import { loginReducer } from "../../reducer/login-reducer";
 import { useToast } from "hooks/useToast";
+import { loginHandler, guestHandler } from "services/auth";
 export const Login = () => {
   const [{ email, password, passwordType }, dispatch] = useReducer(
     loginReducer,
@@ -19,54 +20,15 @@ export const Login = () => {
   const navigate = useNavigate();
   const location = useLocation();
   const from = location.state?.from?.pathname || "/";
-  const loginHandler = async (e) => {
-    e.preventDefault();
-
-    try {
-      const response = await axios.post("/api/auth/login", {
-        email,
-        password,
-      });
-      const {
-        status,
-        data: { encodedToken, foundUser },
-      } = response;
-      if (status >= 200 && status <= 299) {
-        setAuth({ ...auth, auth: true, user: foundUser, token: encodedToken });
-        localStorage.setItem("token", encodedToken);
-        navigate(from, { replace: true });
-        showToast("success", "Successfully Logged In!");
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
-  const guestHandler = async (e) => {
-    e.preventDefault();
-    try {
-      const response = await axios.post("/api/auth/login", {
-        email: "Testuser@gmail.com",
-        password: "Testuser",
-      });
-      const {
-        status,
-        data: { encodedToken, foundUser },
-      } = response;
-      if (status >= 200 && status <= 299) {
-        setAuth({ ...auth, auth: true, user: foundUser, token: encodedToken });
-        localStorage.setItem("token", encodedToken);
-        navigate(from, { replace: true });
-        showToast("success", "Successfully Logged In!");
-      }
-    } catch (error) {
-      console.log("something went wrong");
-    }
-  };
 
   return (
     <div className="login-form">
       <h2 className="login-title">login</h2>
-      <form onSubmit={loginHandler}>
+      <form
+        onSubmit={(e) =>
+          loginHandler(e, axios, setAuth, auth, navigate, showToast, from, email, password)
+        }
+      >
         <label htmlFor="email">
           Email
           <input
@@ -119,10 +81,13 @@ export const Login = () => {
           </div>
         </label>
 
-        <button className="login-btn">
-          login
-        </button>
-        <button className="login-btn" onClick={guestHandler}>
+        <button className="login-btn">login</button>
+        <button
+          className="login-btn"
+          onClick={(e) =>
+            guestHandler(e, axios, setAuth, auth, navigate, showToast, from)
+          }
+        >
           login with guest credentials
         </button>
         <div className="new-account">
